[RCM] Abstract how RC agent clients retrieve config updates
The "agent client" implementation was hard-coded to only work over a
secure GRPC client provided by a utility package in the datadog core
agent. We have a use case coming up where we want to be able to run both
the core remote config service AND an agent client in the same process
without having to setup and run the secure GRPC server within that
process.

In order to do this, we define an interface for the main config update
request and pull out the GRPC logic into a type that implements this
interface. We provide a convenience constructor to create an RC agent
client backed by the secure GRPC client to prevent the GRPC constructor
from needing to leave the internal package.